Bulk of changes
Production ready flask app accomodating minified react app. useActionsColumn hook created as it was inapproriately big in actions.js file. NavBar is created from Bootstrap. UserRole hook is created providing userInfo. Issue: Userinfo not receiving before any rendering
