+++ react_frontend/package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.2.2",
     "bootstrap": "^5.2.3",
     "dateformat": "^5.0.3",
     "http-proxy-middleware": "^2.0.6",
