+++ react_frontend/src/App.js
@@ -5,15 +5,6 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { RelvalTable } from './features/relvals/RelvalTable';
 
 function App() {
-  React.useEffect(() =>{
-    fetch('/api/search?db_name=tickets', 
-      { mode: 'cors',
-        headers: {"Access-Control-Allow-Origin": "*"}
-      }
-    )
-    .then(res => res.json())
-    .then(data => console.log(data))
-  }, []);
 
   return (
     <BrowserRouter>
