+++ react_frontend/src/features/components/NavBar.css
@@ -1,0 +1,36 @@
+/*For laoding spinner to stay on top of footer*/
+footer.fixed-bottom {
+  z-index: 1028;
+}
+
+/* Highlight navbar buttons on hover */
+.mynav-item {
+  /*padding-right: 5px;*/
+  padding-left: 5px;
+}
+.mynav-link {
+  display: flex;
+  /*padding: .5rem 1rem;*/
+  padding-top: 5px;
+  padding-bottom: 5px;
+  padding-right: 10px;
+  padding-left: 10px; 
+  color:  black;
+}
+.mynav-link:hover{
+  color: black;
+  opacity: 1;
+  border-radius:5px;
+  /*font-weight: bold;*/
+  background: lavender;
+}
+
+.navbar-right {
+  margin-left: auto;
+  margin-right: 0px;
+}
+
+div.container {
+  margin-left: 0%;
+  margin-right: 0%;
+}
